requirements.txt
asgiref==3.5.2
Django==4.0.6
djangorestframework==3.13.1
drf-nested-routers==0.93.4
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.1
qrcode==7.3.1
sqlparse==0.4.2
